Add group and connection management for messaging, enhance PresenceHub and PresenceTracker for user connection updates

+++ API/SignalR/MessageHub.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.SignalR;
 namespace API.SignalR;
 
 public class MessageHub(IMessageRepository messageRepository, IUserRepository userRepository, 
-IMapper mapper) : Hub
+IMapper mapper, IHubContext<PresenceHub> presenceHub) : Hub
 {
     public override async Task OnConnectedAsync()
     {
@@ -19,16 +19,19 @@ IMapper mapper) : Hub
             throw new Exception("Cannot join group");
         var groupName = GetGroupName(Context.User.GetUserName(), otherUser);
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
         var messages = await messageRepository.GetMessageThread(Context.User.GetUserName(), otherUser!);
 
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
     }
 
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageRoup();
+        var group = await RemoveFromMessageRoup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
 
@@ -61,6 +64,15 @@ IMapper mapper) : Hub
             {
                 message.DateRead = DateTime.UtcNow;
             }
+            else
+            {
+                var connections = await PresenceTracker.GetConnectionsForUser(recipient.UserName);
+                if(connections != null && connections?.Count != null)
+                {
+                    await presenceHub.Clients.Clients(connections).SendAsync("NewMessageReceived", 
+                    new{username = sender.UserName, KnownAs = sender.KnownAs});
+                }
+            }
 
             messageRepository.AddMessage(message);
 
@@ -70,7 +82,7 @@ IMapper mapper) : Hub
             }
     }
 
-    private async Task<bool> AddToGroup(string groupName)
+    private async Task<Group> AddToGroup(string groupName)
     {
         var username = Context.User?.GetUserName() ?? throw new Exception("Could not get username");
         var group = await messageRepository.GetMessageGroup(groupName);
@@ -84,17 +96,22 @@ IMapper mapper) : Hub
 
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if(await messageRepository.SaveAllAsync()) return group;
+
+        throw new HubException("Failed to join group");
     }
 
-    private async Task RemoveFromMessageRoup()
+    private async Task<Group> RemoveFromMessageRoup()
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if(connection != null)
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+        if(connection != null && group != null)
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if(await messageRepository.SaveAllAsync()) return group;
         }
+        
+        throw new Exception("Failed to  remove from group");
     }
 
     private string GetGroupName(string caller, string? other)
